requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.0.1
waitress==2.1.2